+++ src/hr/avrbanac/openglplayground/loaders/ModelLoader.java
@@ -45,6 +45,14 @@ public class ModelLoader {
         return new RawModel(vaoID, indices.length);
     }
     
+    public RawModel loadToVAO(float[] positions) {
+        int vaoID = createVAO();
+        storeDataInAttributeList(0, 2, positions);
+        unbindVAO();
+        
+        return new RawModel(vaoID, positions.length/2);
+    }
+    
     private int createVAO() {
         int vaoID = GL30.glGenVertexArrays();
         vaos.add(vaoID);